device_connector/requirements.txt
paho_mqtt==1.6.1
pydantic==2.10.4
python-dotenv==1.0.1
Requests==2.32.3
